Move date_deleted field to plan table
