+++ src/main/java/org/opensrp/domain/postgres/Plan.java
@@ -6,7 +6,7 @@ public class Plan {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database column core.plan.id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     private String id;
 
@@ -15,16 +15,25 @@ public class Plan {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database column core.plan.json
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     private Object json;
 
     /**
      *
      * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column core.plan.date_deleted
+     *
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
+     */
+    private Long dateDeleted;
+
+    /**
+     *
+     * This field was generated by MyBatis Generator.
      * This field corresponds to the database column core.plan.server_version
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     private Long serverVersion;
 
@@ -34,7 +43,7 @@ public class Plan {
      *
      * @return the value of core.plan.id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public String getId() {
         return id;
@@ -46,7 +55,7 @@ public class Plan {
      *
      * @param id the value for core.plan.id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setId(String id) {
         this.id = id;
@@ -58,7 +67,7 @@ public class Plan {
      *
      * @return the value of core.plan.json
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public Object getJson() {
         return json;
@@ -70,10 +79,34 @@ public class Plan {
      *
      * @param json the value for core.plan.json
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setJson(Object json) {
         this.json = json;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column core.plan.date_deleted
+     *
+     * @return the value of core.plan.date_deleted
+     *
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
+     */
+    public Long getDateDeleted() {
+        return dateDeleted;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column core.plan.date_deleted
+     *
+     * @param dateDeleted the value for core.plan.date_deleted
+     *
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
+     */
+    public void setDateDeleted(Long dateDeleted) {
+        this.dateDeleted = dateDeleted;
     }
 
     /**
@@ -82,7 +115,7 @@ public class Plan {
      *
      * @return the value of core.plan.server_version
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public Long getServerVersion() {
         return serverVersion;
@@ -94,7 +127,7 @@ public class Plan {
      *
      * @param serverVersion the value for core.plan.server_version
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setServerVersion(Long serverVersion) {
         this.serverVersion = serverVersion;
