--- src/main/java/org/opensrp/domain/postgres/PlanMetadata.java
@@ -1,36 +1,0 @@
-package org.opensrp.domain.postgres;
-
-public class PlanMetadata extends PlanMetadataKey {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column core.plan_metadata.date_deleted
-     *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
-     */
-    private Long dateDeleted;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column core.plan_metadata.date_deleted
-     *
-     * @return the value of core.plan_metadata.date_deleted
-     *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
-     */
-    public Long getDateDeleted() {
-        return dateDeleted;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column core.plan_metadata.date_deleted
-     *
-     * @param dateDeleted the value for core.plan_metadata.date_deleted
-     *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
-     */
-    public void setDateDeleted(Long dateDeleted) {
-        this.dateDeleted = dateDeleted;
-    }
-}
