+++ src/main/java/org/opensrp/domain/postgres/PlanMetadataExample.java
@@ -8,7 +8,7 @@ public class PlanMetadataExample {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     protected String orderByClause;
 
@@ -16,7 +16,7 @@ public class PlanMetadataExample {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     protected boolean distinct;
 
@@ -24,7 +24,7 @@ public class PlanMetadataExample {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     protected List<Criteria> oredCriteria;
 
@@ -32,7 +32,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public PlanMetadataExample() {
         oredCriteria = new ArrayList<Criteria>();
@@ -42,7 +42,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setOrderByClause(String orderByClause) {
         this.orderByClause = orderByClause;
@@ -52,7 +52,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public String getOrderByClause() {
         return orderByClause;
@@ -62,7 +62,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setDistinct(boolean distinct) {
         this.distinct = distinct;
@@ -72,7 +72,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public boolean isDistinct() {
         return distinct;
@@ -82,7 +82,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public List<Criteria> getOredCriteria() {
         return oredCriteria;
@@ -92,7 +92,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void or(Criteria criteria) {
         oredCriteria.add(criteria);
@@ -102,7 +102,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public Criteria or() {
         Criteria criteria = createCriteriaInternal();
@@ -114,7 +114,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public Criteria createCriteria() {
         Criteria criteria = createCriteriaInternal();
@@ -128,7 +128,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     protected Criteria createCriteriaInternal() {
         Criteria criteria = new Criteria();
@@ -139,7 +139,7 @@ public class PlanMetadataExample {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void clear() {
         oredCriteria.clear();
@@ -151,7 +151,7 @@ public class PlanMetadataExample {
      * This class was generated by MyBatis Generator.
      * This class corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     protected abstract static class GeneratedCriteria {
         protected List<Criterion> criteria;
@@ -333,73 +333,13 @@ public class PlanMetadataExample {
             addCriterion("plan_id not between", value1, value2, "planId");
             return (Criteria) this;
         }
-
-        public Criteria andDateDeletedIsNull() {
-            addCriterion("date_deleted is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedIsNotNull() {
-            addCriterion("date_deleted is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedEqualTo(Long value) {
-            addCriterion("date_deleted =", value, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedNotEqualTo(Long value) {
-            addCriterion("date_deleted <>", value, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedGreaterThan(Long value) {
-            addCriterion("date_deleted >", value, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedGreaterThanOrEqualTo(Long value) {
-            addCriterion("date_deleted >=", value, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedLessThan(Long value) {
-            addCriterion("date_deleted <", value, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedLessThanOrEqualTo(Long value) {
-            addCriterion("date_deleted <=", value, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedIn(List<Long> values) {
-            addCriterion("date_deleted in", values, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedNotIn(List<Long> values) {
-            addCriterion("date_deleted not in", values, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedBetween(Long value1, Long value2) {
-            addCriterion("date_deleted between", value1, value2, "dateDeleted");
-            return (Criteria) this;
-        }
-
-        public Criteria andDateDeletedNotBetween(Long value1, Long value2) {
-            addCriterion("date_deleted not between", value1, value2, "dateDeleted");
-            return (Criteria) this;
-        }
     }
 
     /**
      * This class was generated by MyBatis Generator.
      * This class corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated do_not_delete_during_merge Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated do_not_delete_during_merge Thu May 02 10:52:55 EAT 2019
      */
     public static class Criteria extends GeneratedCriteria {
 
@@ -412,7 +352,7 @@ public class PlanMetadataExample {
      * This class was generated by MyBatis Generator.
      * This class corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public static class Criterion {
         private String condition;
