+++ src/main/java/org/opensrp/domain/postgres/PlanMetadataKey.java
@@ -6,7 +6,7 @@ public class PlanMetadataKey {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database column core.plan_metadata.operation_area_id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     private String operationAreaId;
 
@@ -15,7 +15,7 @@ public class PlanMetadataKey {
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database column core.plan_metadata.plan_id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     private String planId;
 
@@ -25,7 +25,7 @@ public class PlanMetadataKey {
      *
      * @return the value of core.plan_metadata.operation_area_id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public String getOperationAreaId() {
         return operationAreaId;
@@ -37,7 +37,7 @@ public class PlanMetadataKey {
      *
      * @param operationAreaId the value for core.plan_metadata.operation_area_id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setOperationAreaId(String operationAreaId) {
         this.operationAreaId = operationAreaId;
@@ -49,7 +49,7 @@ public class PlanMetadataKey {
      *
      * @return the value of core.plan_metadata.plan_id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public String getPlanId() {
         return planId;
@@ -61,7 +61,7 @@ public class PlanMetadataKey {
      *
      * @param planId the value for core.plan_metadata.plan_id
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     public void setPlanId(String planId) {
         this.planId = planId;
