+++ src/main/java/org/opensrp/repository/postgres/mapper/PlanMapper.java
@@ -10,7 +10,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     long countByExample(PlanExample example);
 
@@ -18,7 +18,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int deleteByExample(PlanExample example);
 
@@ -26,7 +26,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int deleteByPrimaryKey(String id);
 
@@ -34,7 +34,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int insert(Plan record);
 
@@ -42,7 +42,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int insertSelective(Plan record);
 
@@ -50,7 +50,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     List<Plan> selectByExample(PlanExample example);
 
@@ -58,7 +58,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     Plan selectByPrimaryKey(String id);
 
@@ -66,7 +66,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int updateByExampleSelective(@Param("record") Plan record, @Param("example") PlanExample example);
 
@@ -74,7 +74,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int updateByExample(@Param("record") Plan record, @Param("example") PlanExample example);
 
@@ -82,7 +82,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int updateByPrimaryKeySelective(Plan record);
 
@@ -90,7 +90,7 @@ public interface PlanMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int updateByPrimaryKey(Plan record);
 }
