+++ src/main/java/org/opensrp/repository/postgres/mapper/PlanMetadataMapper.java
@@ -2,7 +2,6 @@ package org.opensrp.repository.postgres.mapper;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
-import org.opensrp.domain.postgres.PlanMetadata;
 import org.opensrp.domain.postgres.PlanMetadataExample;
 import org.opensrp.domain.postgres.PlanMetadataKey;
 
@@ -11,7 +10,7 @@ public interface PlanMetadataMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     long countByExample(PlanMetadataExample example);
 
@@ -19,7 +18,7 @@ public interface PlanMetadataMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int deleteByExample(PlanMetadataExample example);
 
@@ -27,7 +26,7 @@ public interface PlanMetadataMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
     int deleteByPrimaryKey(PlanMetadataKey key);
 
@@ -35,63 +34,39 @@ public interface PlanMetadataMapper {
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
-    int insert(PlanMetadata record);
+    int insert(PlanMetadataKey record);
 
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
-    int insertSelective(PlanMetadata record);
+    int insertSelective(PlanMetadataKey record);
 
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
-    List<PlanMetadata> selectByExample(PlanMetadataExample example);
+    List<PlanMetadataKey> selectByExample(PlanMetadataExample example);
 
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
-    PlanMetadata selectByPrimaryKey(PlanMetadataKey key);
+    int updateByExampleSelective(@Param("record") PlanMetadataKey record, @Param("example") PlanMetadataExample example);
 
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table core.plan_metadata
      *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
+     * @mbg.generated Thu May 02 10:52:55 EAT 2019
      */
-    int updateByExampleSelective(@Param("record") PlanMetadata record, @Param("example") PlanMetadataExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table core.plan_metadata
-     *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
-     */
-    int updateByExample(@Param("record") PlanMetadata record, @Param("example") PlanMetadataExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table core.plan_metadata
-     *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
-     */
-    int updateByPrimaryKeySelective(PlanMetadata record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table core.plan_metadata
-     *
-     * @mbg.generated Tue Apr 30 21:54:17 EAT 2019
-     */
-    int updateByPrimaryKey(PlanMetadata record);
+    int updateByExample(@Param("record") PlanMetadataKey record, @Param("example") PlanMetadataExample example);
 }
