+++ src/main/java/org/opensrp/repository/postgres/mapper/xml/PlanMapper.xml
@@ -5,17 +5,18 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     <id column="id" jdbcType="VARCHAR" property="id" />
     <result column="json" jdbcType="OTHER" property="json" />
+    <result column="date_deleted" jdbcType="BIGINT" property="dateDeleted" />
     <result column="server_version" jdbcType="BIGINT" property="serverVersion" />
   </resultMap>
   <sql id="Example_Where_Clause">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     <where>
       <foreach collection="oredCriteria" item="criteria" separator="or">
@@ -49,7 +50,7 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     <where>
       <foreach collection="example.oredCriteria" item="criteria" separator="or">
@@ -83,15 +84,15 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
-    id, json, server_version
+    id, json, date_deleted, server_version
   </sql>
   <select id="selectByExample" parameterType="org.opensrp.domain.postgres.PlanExample" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     select
     <if test="distinct">
@@ -110,7 +111,7 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     select 
     <include refid="Base_Column_List" />
@@ -121,7 +122,7 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     delete from core.plan
     where id = #{id,jdbcType=VARCHAR}
@@ -130,7 +131,7 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     delete from core.plan
     <if test="_parameter != null">
@@ -141,18 +142,18 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
-    insert into core.plan (id, json, server_version
-      )
-    values (#{id,jdbcType=VARCHAR}, #{json,jdbcType=OTHER}, #{serverVersion,jdbcType=BIGINT}
-      )
+    insert into core.plan (id, json, date_deleted, 
+      server_version)
+    values (#{id,jdbcType=VARCHAR}, #{json,jdbcType=OTHER}, #{dateDeleted,jdbcType=BIGINT}, 
+      #{serverVersion,jdbcType=BIGINT})
   </insert>
   <insert id="insertSelective" parameterType="org.opensrp.domain.postgres.Plan">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     insert into core.plan
     <trim prefix="(" suffix=")" suffixOverrides=",">
@@ -161,6 +162,9 @@
       </if>
       <if test="json != null">
         json,
+      </if>
+      <if test="dateDeleted != null">
+        date_deleted,
       </if>
       <if test="serverVersion != null">
         server_version,
@@ -173,6 +177,9 @@
       <if test="json != null">
         #{json,jdbcType=OTHER},
       </if>
+      <if test="dateDeleted != null">
+        #{dateDeleted,jdbcType=BIGINT},
+      </if>
       <if test="serverVersion != null">
         #{serverVersion,jdbcType=BIGINT},
       </if>
@@ -182,7 +189,7 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     select count(*) from core.plan
     <if test="_parameter != null">
@@ -193,7 +200,7 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     update core.plan
     <set>
@@ -202,6 +209,9 @@
       </if>
       <if test="record.json != null">
         json = #{record.json,jdbcType=OTHER},
+      </if>
+      <if test="record.dateDeleted != null">
+        date_deleted = #{record.dateDeleted,jdbcType=BIGINT},
       </if>
       <if test="record.serverVersion != null">
         server_version = #{record.serverVersion,jdbcType=BIGINT},
@@ -215,11 +225,12 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     update core.plan
     set id = #{record.id,jdbcType=VARCHAR},
       json = #{record.json,jdbcType=OTHER},
+      date_deleted = #{record.dateDeleted,jdbcType=BIGINT},
       server_version = #{record.serverVersion,jdbcType=BIGINT}
     <if test="_parameter != null">
       <include refid="Update_By_Example_Where_Clause" />
@@ -229,12 +240,15 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     update core.plan
     <set>
       <if test="json != null">
         json = #{json,jdbcType=OTHER},
+      </if>
+      <if test="dateDeleted != null">
+        date_deleted = #{dateDeleted,jdbcType=BIGINT},
       </if>
       <if test="serverVersion != null">
         server_version = #{serverVersion,jdbcType=BIGINT},
@@ -246,10 +260,11 @@
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Tue Apr 30 21:54:17 EAT 2019.
+      This element was generated on Thu May 02 10:52:55 EAT 2019.
     -->
     update core.plan
     set json = #{json,jdbcType=OTHER},
+      date_deleted = #{dateDeleted,jdbcType=BIGINT},
       server_version = #{serverVersion,jdbcType=BIGINT}
     where id = #{id,jdbcType=VARCHAR}
   </update>
